Move person permissions classes into person/permissions
