+++ src/test/java/org/synyx/urlaubsverwaltung/person/permissions/PersonPermissionsDtoValidatorTest.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.person.web;
+package org.synyx.urlaubsverwaltung.person.permissions;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -7,6 +7,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.validation.Errors;
 import org.synyx.urlaubsverwaltung.application.application.Application;
+import org.synyx.urlaubsverwaltung.person.permissions.PersonPermissionsDto;
+import org.synyx.urlaubsverwaltung.person.permissions.PersonPermissionsDtoValidator;
 
 import java.util.ArrayList;
