+++ src/test/java/org/synyx/urlaubsverwaltung/person/permissions/PersonPermissionsViewControllerTest.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.person.web;
+package org.synyx.urlaubsverwaltung.person.permissions;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -13,6 +13,8 @@ import org.synyx.urlaubsverwaltung.department.DepartmentService;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.person.PersonService;
 import org.synyx.urlaubsverwaltung.person.UnknownPersonException;
+import org.synyx.urlaubsverwaltung.person.permissions.PersonPermissionsDtoValidator;
+import org.synyx.urlaubsverwaltung.person.permissions.PersonPermissionsViewController;
 import org.synyx.urlaubsverwaltung.security.SessionService;
 
 import java.util.List;
